Namespace serializers under Api::Internal::V1
This will allow us to render different versions
of JSON on different version endpoints.
