+++ app/serializers/api/internal/v1/hearing_event_serializer.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Api
+  module Internal
+    module V1
+      class HearingEventSerializer
+        include JSONAPI::Serializer
+        set_type :hearing_events
+
+        attributes :description
+        attributes :occurred_at
+      end
+    end
+  end
+end
