+++ app/serializers/api/internal/v1/hearing_summary_serializer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Api
+  module Internal
+    module V1
+      class HearingSummarySerializer
+        include JSONAPI::Serializer
+        set_type :hearing_summaries
+
+        attributes :hearing_type, :hearing_days
+      end
+    end
+  end
+end
