+++ app/serializers/api/internal/v1/provider_serializer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Api
+  module Internal
+    module V1
+      class ProviderSerializer
+        include JSONAPI::Serializer
+        set_type :providers
+
+        attributes :name, :role
+      end
+    end
+  end
+end
