--- app/serializers/cracked_ineffective_trial_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class CrackedIneffectiveTrialSerializer
-  include JSONAPI::Serializer
-  set_type :cracked_ineffective_trial
-
-  attributes :id, :code, :type, :description
-end
