--- app/serializers/hearing_event_serializer.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class HearingEventSerializer
-  include JSONAPI::Serializer
-  set_type :hearing_events
-
-  attributes :description
-  attributes :occurred_at
-end
