--- app/serializers/hearing_summary_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class HearingSummarySerializer
-  include JSONAPI::Serializer
-  set_type :hearing_summaries
-
-  attributes :hearing_type, :hearing_days
-end
