--- app/serializers/provider_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class ProviderSerializer
-  include JSONAPI::Serializer
-  set_type :providers
-
-  attributes :name, :role
-end
