+++ spec/serializers/api/internal/v1/hearing_summary_serializer_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe HearingSummarySerializer do
+RSpec.describe Api::Internal::V1::HearingSummarySerializer do
   subject { described_class.new(hearing_summary).serializable_hash }
 
   let(:hearing_summary) do
